+++ typings/helpers/roundNumber.d.ts
@@ -1,4 +1,3 @@
-import { BigNumber } from "bignumber.js";
 import { RoundingMode } from "../typing";
 type RoundingOptions = {
     roundMode: RoundingMode;